Rename method buttonSignInDisplayed to buttonSignUpDisplayed Move code from class NotValidLogin.java to LoginTest and removed not usable classes
