--- oliinykProj/src/test/java/loginTest/LoginAndPasswordEnter.java
@@ -1,16 +1,0 @@
-package loginTest;
-
-import java.util.Scanner;
-
-public class LoginAndPasswordEnter {
-
-    public static String loginEnter(){
-        Scanner scanner = new Scanner(System.in);
-        return scanner.nextLine();
-    }
-
-    public static String passEnter(){
-        Scanner scanner = new Scanner(System.in);
-        return scanner.nextLine();
-    }
-}
