+++ oliinykProj/src/test/java/loginTest/LoginTest.java
@@ -35,6 +35,27 @@ public class LoginTest {
 
         Assert.assertTrue("Button Sign Out not displayed", dispayedButtonSignOut());
 
+        driver.findElement(By.xpath(".//button[text()='Sign Out']")).click();
+        Assert.assertTrue("Button 'Sign up for OurApp' not displayed",buttonSignUpDisplayed());
+        System.out.println("Log out and enter for main page");
+
+
+        driver.findElement(By.xpath(".//input[@name='username' and @placeholder='Username']")).clear();
+        driver.findElement(By.xpath(".//input[@name='username' and @placeholder='Username']")).sendKeys("abcdefg");
+        System.out.println("Login was entered");
+
+        driver.findElement(By.xpath(".//input[@placeholder='Password']")).clear();
+        driver.findElement(By.xpath(".//input[@placeholder='Password']")).sendKeys("123456qwert");
+        System.out.println("Password was entered");
+
+        driver.findElement(By.xpath(".//button[text()='Sign In']")).click();
+        System.out.println("Sign in button was clicked");
+
+        Assert.assertTrue("Button 'Sign up for OurApp' not displayed",buttonSignUpDisplayed());
+        System.out.println("Button 'Sign up for OurApp' is displayed");
+        Assert.assertTrue("Wrong text is displayed",textErrorIsDisplayed());
+        System.out.println("Text is right");
+
         driver.quit();
         System.out.println("Browser closed");
 
@@ -43,6 +64,23 @@ public class LoginTest {
     private boolean dispayedButtonSignOut(){
         try {
             return driver.findElement(By.xpath(".//button[text()='Sign Out']")).isDisplayed();
+        }catch (Exception e){
+            return false;
+        }
+    }
+
+    //proverka vidimosti knopki registracii
+    private boolean buttonSignUpDisplayed(){
+        try {
+            return driver.findElement(By.xpath(".//button[@type='submit']")).isDisplayed();
+        }catch (Exception e){
+            return false;
+        }
+    }
+    //proverka vivedennogo texta posle ne korrektnogo logina
+    private boolean textErrorIsDisplayed(){
+        try {
+            return driver.findElement(By.xpath(".//div[@class='alert alert-danger text-center']")).getText().equalsIgnoreCase("error");
         }catch (Exception e){
             return false;
         }
